+++ frontend/eslint-plugin-ddangkong/enforce-is-boolean.js
